projects/MCP server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-groq>=0.3.8",
    "langchain-mcp-adapters>=0.1.11",
    "langgraph>=0.6.10",
]
